move event type to separate file outside of pipeline; rename events to generalize to other schedulers than just the pipeline
